// src/booking/enums/booking-activity-type.enum.ts
export enum BookingActivityType {
  CREATED = "CREATED",                 
  ACCEPTED = "ACCEPTED",               
  ASSIGNED = "ASSIGNED",               
  WORKER_ACCEPTED = "WORKER_ACCEPTED",
  WORKER_REJECTED = "WORKER_REJECTED",
  WORKER_STARTED = "WORKER_STARTED",
  WORKER_COMPLETED = "WORKER_COMPLETED",
  ALL_WORKERS_STARTED = "ALL_WORKERS_STARTED",
  ALL_WORKERS_COMPLETED = "ALL_WORKERS_COMPLETED",
  CANCELLED_BY_CUSTOMER = "CANCELLED_BY_CUSTOMER",
  CANCELLED_BY_WORKER = "CANCELLED_BY_WORKER",
  CANCELLED_BY_PLATFORM = "CANCELLED_BY_PLATFORM",
  NO_SHOW = "NO_SHOW",
  FINALIZED = "FINALIZED", 
  PARTIALLY_PAID = "PARTIALLY_PAID",             
  PAID = "PAID",
  REFUNDED = "REFUNDED",
  REVIEWED = "REVIEWED",
  EXPIRED = "EXPIRED"
}